requirements.txt
openai>=1.0.0
python-dotenv
pytest>=7.0.0